Do not make matplotlib a requirement for the package, just for the scripts
