+++ requirements.txt
@@ -14,11 +14,8 @@ importlib-metadata==0.23
 ipython==7.6.1
 ipython-genutils==0.2.0
 jedi==0.14.0
-kiwisolver==1.1.0
-matplotlib==3.1.1
 more-itertools==7.2.0
 nodeenv==1.3.3
-numpy==1.16.4
 parso==0.5.0
 pexpect==4.7.0
 pickleshare==0.7.5
@@ -28,8 +25,6 @@ prompt-toolkit==2.0.9
 ptyprocess==0.6.0
 Pygments==2.4.2
 pyparsing==2.4.0
-PyQt5==5.13.1
-PyQt5-sip==12.7.0
 python-bugzilla==2.3.0
 python-dateutil==2.8.0
 PyYAML==5.1.1
