+++ setup.cfg
@@ -41,11 +41,8 @@ install_requires =
     ipython
     ipython-genutils
     jedi
-    kiwisolver
-    matplotlib
     more-itertools
     nodeenv
-    numpy
     parso
     pexpect
     pickleshare
@@ -56,7 +53,6 @@ install_requires =
     pyparsing
     python-bugzilla
     python-dateutil
-    PyQt5
     PyYAML
     requests
     six
